Move tests back to test_spark.py, add comments
